converted result to double

+++ src/BmiService.java
@@ -1,8 +1,8 @@
 public class BmiService {
 
-    public int calculate(double height, double weight) {
+    public double calculate(double height, double weight) {
         double bodymassindex;
         bodymassindex = weight / (height * height);
-        return (int) bodymassindex;
+        return bodymassindex;
     }
 }
